refactor: rename agents to turns in agent_harness
