+++ core/agent_harness/__init__.py
@@ -3,7 +3,7 @@
 This package owns the surface-agnostic turn harness around the shared
 ``core.agent.Agent`` loop. It was extracted out of ``interactive_shell`` so the
 same harness can drive the interactive terminal **and** be executed headlessly via a plain API call
-(:func:`core.agent_harness.agents.headless_agent.dispatch_message_to_headless_agent`).
+(:func:`core.agent_harness.turns.headless_dispatch.dispatch_message_to_headless_agent`).
 
 Hard boundary: nothing under ``agent_harness/`` may import from
 ``interactive_shell``. The dependency direction is one-way:
@@ -12,14 +12,6 @@ Hard boundary: nothing under ``agent_harness/`` may import from
 
 from __future__ import annotations
 
-from core.agent_harness.agents.action_agent import ToolCallingDeps
-from core.agent_harness.agents.action_agent import (
-    run_action_agent_turn as execute_action_agent_turn,
-)
-from core.agent_harness.agents.evidence_agent import gather_tool_evidence
-from core.agent_harness.agents.evidence_agent import gather_tool_evidence as gather_evidence
-from core.agent_harness.agents.headless_agent import dispatch_message_to_headless_agent
-from core.agent_harness.agents.turn_orchestrator import run_turn, stream_answer
 from core.agent_harness.harness import AgentHarness, HarnessConfig, HarnessStartupResult
 from core.agent_harness.models.turn_results import ShellTurnResult, ToolCallingTurnResult
 from core.agent_harness.models.turn_snapshot import (
@@ -27,6 +19,14 @@ from core.agent_harness.models.turn_snapshot import (
     TurnSnapshot,
     TurnSnapshotSource,
 )
+from core.agent_harness.turns.action_driver import ToolCallingDeps
+from core.agent_harness.turns.action_driver import (
+    run_action_agent_turn as execute_action_agent_turn,
+)
+from core.agent_harness.turns.evidence_driver import gather_tool_evidence
+from core.agent_harness.turns.evidence_driver import gather_tool_evidence as gather_evidence
+from core.agent_harness.turns.headless_dispatch import dispatch_message_to_headless_agent
+from core.agent_harness.turns.orchestrator import run_turn, stream_answer
 
 __all__ = [
     "AgentHarness",
