+++ gateway/AGENTS.md
@@ -27,7 +27,7 @@ Pytest discovers these tests through `pytest.ini`; scoped CI maps changes under
 
 - **No persistent gateway `Agent` instance.** Each inbound message gets a
   per-chat `Session` from `SessionResolver` and is handled by the shared
-  headless dispatch path (`core.agent_harness.agents.headless_agent`).
+  headless dispatch path (`core.agent_harness.turns.headless_dispatch`).
 - The turn handler callback signature is exactly four arguments: `text`,
   `session`, `sink`, and `logger`. Do not reintroduce `chat_id` into this
   contract; the sink owns chat transport details.
