+++ gateway/tests/main/test_antartica_slack.py
@@ -18,11 +18,11 @@ from unittest.mock import MagicMock
 import pytest
 from rich.console import Console
 
-from core.agent_harness.agents.action_agent import ToolCallingDeps, run_action_agent_turn
 from core.agent_harness.providers.default_providers import DefaultToolProvider
 from core.agent_harness.session import Session
 from core.agent_harness.session.storage.memory import InMemorySessionStorage
 from core.agent_harness.tools.action_tools import action_tool_names
+from core.agent_harness.turns.action_driver import ToolCallingDeps, run_action_agent_turn
 from core.llm.types import AgentLLMResponse, ToolCall
 from tools.registry import clear_tool_registry_cache
