+++ surfaces/interactive_shell/runtime/integration_tool_gathering.py
@@ -1,7 +1,7 @@
 """Gather integration evidence for a conversational shell answer.
 
 The bounded think -> call-tools -> observe loop lives in the decoupled
-:func:`core.agent_harness.agents.evidence_agent.gather_tool_evidence`. This module is the terminal adapter:
+:func:`core.agent_harness.turns.evidence_driver.gather_tool_evidence`. This module is the terminal adapter:
 it renders each gathering step to the console and persists the gathered tool
 calls into the shell's session storage, then hands the collected observation back
 to :func:`interactive_shell.runtime.answer_turn.answer_shell_question`.
@@ -16,9 +16,9 @@ from typing import Any
 from rich.console import Console
 from rich.markup import escape
 
-from core.agent_harness.agents import evidence_agent
-from core.agent_harness.agents.evidence_agent import EvidenceAgentFactory
 from core.agent_harness.session import Session
+from core.agent_harness.turns import evidence_driver
+from core.agent_harness.turns.evidence_driver import EvidenceAgentFactory
 from surfaces.interactive_shell.ui import DIM
 from surfaces.interactive_shell.utils.error_handling.exception_reporting import report_exception
 from surfaces.shared.tool_labels import tool_short_label, tool_source_label
@@ -104,7 +104,7 @@ def _format_gathering_progress_line(
 
 def _resolve_gather_integrations(session: Session, message: str) -> dict[str, Any]:
     """Resolve gather integrations through the decoupled agent helper."""
-    return evidence_agent._resolve_gather_integrations(session, message)  # noqa: SLF001
+    return evidence_driver._resolve_gather_integrations(session, message)  # noqa: SLF001
 
 
 def _truncate(text: str, limit: int) -> str:
@@ -173,7 +173,7 @@ def gather_integration_tool_evidence(
     def persist(executed: list[tuple[Any, Any]]) -> None:
         _persist_tool_calls(session, executed)
 
-    return evidence_agent.gather_tool_evidence(
+    return evidence_driver.gather_tool_evidence(
         message,
         session,
         on_progress=on_progress,
