+++ tests/core/agent/orchestration/test_cross_surface_components.py
@@ -9,10 +9,10 @@ from unittest.mock import MagicMock
 import pytest
 from rich.console import Console
 
-from core.agent_harness.agents.turn_orchestrator import run_turn
 from core.agent_harness.models.turn_results import ShellTurnResult, ToolCallingTurnResult
 from core.agent_harness.providers.default_providers import DefaultToolProvider
 from core.agent_harness.session import InMemorySessionStorage, Session
+from core.agent_harness.turns.orchestrator import run_turn
 from gateway.turn_handler import build_gateway_turn_handler
 
 
@@ -115,7 +115,7 @@ def test_run_turn_populates_resolved_integrations_on_turn_snapshot(
     """run_turn resolves integrations once and hands them to the action path on turn_snapshot."""
     resolved = {"github": {"configured": True}}
     monkeypatch.setattr(
-        "core.agent_harness.agents.turn_orchestrator.resolve_and_cache_integrations",
+        "core.agent_harness.turns.orchestrator.resolve_and_cache_integrations",
         lambda _session: resolved,
     )
     captured: list[Any] = []
