+++ tests/core/agent/orchestration/test_model_question_misroute.py
@@ -9,7 +9,7 @@ informational question) is exercised by the live planning scenario
 ``chat_handoff/342-which-model-is-used-now``. These tests pin the *deterministic*
 half: given that the action agent picks ``/model``, the turn ends without
 answering the user, because ``/model`` records no ``last_command_observation`` and
-so the turn router (``core.agent_harness.agents.turn_orchestrator._route_turn``) takes the
+so the turn router (``core.agent_harness.turns.orchestrator._route_turn``) takes the
 ``handled_without_llm`` path instead of summarizing an observation.
 """
