+++ tests/core/agent/test_turn_loop.py
@@ -7,10 +7,10 @@ from typing import Any
 
 from rich.console import Console
 
-from core.agent_harness.agents.turn_orchestrator import run_turn
 from core.agent_harness.providers.default_providers import DefaultTurnAccounting
 from core.agent_harness.session import Session
 from core.agent_harness.session.storage.memory import InMemorySessionStorage
+from core.agent_harness.turns.orchestrator import run_turn
 from surfaces.interactive_shell.runtime.core.state import ReplState, SpinnerState
 from surfaces.interactive_shell.runtime.core.turn_accounting import (
     ToolCallingTurnResult,
