+++ tests/interactive_shell/tools/test_tool_gathering.py
@@ -19,8 +19,8 @@ from rich.console import Console
 import core as runtime_module
 import core.llm.agent_llm_client as agent_llm_client
 import tools.investigation.stages.gather_evidence.tools as investigate_tools
-from core.agent_harness.agents.evidence_agent import EvidenceAgentFactory
 from core.agent_harness.session import Session
+from core.agent_harness.turns.evidence_driver import EvidenceAgentFactory
 from core.llm.types import ToolCall
 from surfaces.interactive_shell.runtime.integration_tool_gathering import (
     _format_gathering_progress_line,
